UPSTREAM: HID: wacom: Rename wacom ID report ID macros
"INTUOSREAD" and "INTUOSWRITE" are poorly named. These are report IDs
for pen ID (proximity) packets. It should be noted that the latter is
only used on Intuos/Intuos2 for a second stylus when DualTrack is in use.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 061099936a724cdd8dd0d69deea5b064b4f7122b)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: I0c69113dea4d823abd7a395bae7a808e50c4e945