UPSTREAM: HID: wacom: Apply lowres quirk to BAMBOO_TOUCH devices
When splitting the touch-only "BAMBOO_TOUCH" type out of the existing
"BAMBOO_PT" type in 3b164a00, the lowres quirk was not updated so that
it would continue to apply to these devices (effectively only the 0xD0).
The absence of this quirk does not significantly impact usability, but
is a correctness issue nonetheless.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit be853fd1c6a5fe127af3392157dfe95e9712a575)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: I8cc5f87470ee1ab3ab3dea645c513dbe71e8a84e