UPSTREAM: net: cdc_ncm: add "ndp_to_end" sysfs attribute
Adding a writable sysfs attribute for the "NDP to end"
quirk flag.

This makes it easier for end users to test new devices for
this firmware bug.  We've been lucky so far, but we should
not depend on reporters capable of rebuilding the driver.

Cc: Enrico Mioso <mrkiko.rs@gmail.com>
Signed-off-by: Bjørn Mork <bjorn@mork.no>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 404814af69d4732276319b90886b81fb2884ae1b)

BUG=b:161488133
TEST=bootperf -o ~/bootdata/atlas-13704-AUE2500CF-01 192.168.86.102 100

Signed-off-by: Grant Grundler <grundler@chromium.org>
Change-Id: Idb1d05694e41c72f7485b8cc1dd9a7569b03c254
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2606988
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>