UPSTREAM: HID: wacom: Fix touchring value reporting
Commit c7f0522 reports incorrect touchring values to userspace. This is
due to its incorrect handling of the 'touched' bit present in the 'ring1'
and 'ring2' variables. Instead of using this bit when determining if a
value should be sent, the ABS_WHEEL and ABS_INPUT check (different?!)
portions of the position bits. Furthermore, the full values of 'ring1'
and 'ring2' are reported to userspace, despite the 'touched' flag
needing to be trimmed beforehand. This commit addresses both issues.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit aaae03e4f7f0c641a258dde855ab677c9b58b155)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: I1d9c1a915ff9dd8b46c1631e0e33b58bff5c74fa