UPSTREAM: platform/chrome: cros_ec_dev - Fix security issue
Prevent memory scribble by checking that ioctl buffer size parameters
are sane.
Without this check, on 32 bits system, if .insize = 0xffffffff - 20 and
.outsize the amount to scribble, we would overflow, allocate a small
amounts and be able to write outside of the malloc'ed area.
Adding a hard limit allows argument checking of the ioctl. With the
current EC, it is expected .insize and .outsize to be at around 512 bytes
or less.

BUG=chromium:583058
TEST=On snappy, check ectool is still working.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Olof Johansson <olof@lixom.net>
(cherry picked from commit 5d749d0bbe811c10d9048cde6dfebc761713abfd)
Change-Id: Ie209fa7f57fc42131dbb3410b2f078312ad94852
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2729625
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>