UPSTREAM: cfg80211: Add option to report the bss entry in connect result
Since cfg80211 maintains separate BSS table entries for APs if the same
BSSID, SSID pair is seen on multiple channels, it is possible that it
can map the current_bss to a BSS entry on the wrong channel. This
current_bss will not get flushed unless disconnected and cfg80211
reports a wrong channel as the associated channel.

Fix this by introducing a new cfg80211_connect_bss() function which is
similar to cfg80211_connect_result(), but it includes an additional
parameter: the bss the STA is connected to. This allows drivers to
provide the exact bss entry that matches the BSS to which the connection
was completed.

Reviewed-by: Jouni Malinen <jouni@qca.qualcomm.com>
Signed-off-by: Vidyullatha Kanchanapally <vkanchan@qti.qualcomm.com>
Signed-off-by: Sunil Dutt <usdutt@qti.qualcomm.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit e705498945ad3a3b945771c5d683df064bb9819c)

BUG=b:175539094
TEST=`tast run $DUT wifi.ChannelHop` on kevin DUT.

Signed-off-by: Yen-lin Lai <yenlinlai@chromium.org>
Change-Id: I672b7506bd687c7ccce086ccae9405ae5515ec68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2703362
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>