UPSTREAM: phy: Add set_mode callback
The initial use for this is for PHYs that have a mode related to USB OTG.
There are several SoCs (e.g. TI OMAP and DA8xx) that have a mode setting
in the USB PHY to override OTG VBUS and ID signals.

Of course, the enum can be expaned in the future to include modes for
other types of PHYs as well.

Suggested-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: David Lechner <david@lechnology.com>
(cherry picked from commit 300eb0139cf27044c67f6005ff17b8434c7843f0)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I8c0290dd10399a626041aa48dd7bec735c324288
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2420039
Reviewed-by: Sean Paul <seanpaul@chromium.org>