UPSTREAM: HID: wacom: leds: handle the switch of the LEDs directly in the kernel
The EKR switches the LED directly, and there is no point in having
userspace handling the switch it self when it's easy enough to do
in the kernel.

The other benefit is that now userspace does not need to have root access
to the LED but need only to read them with user privileges.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Acked-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 34736aa96e313d304ffe0746519833167e66c978)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: I2486e62c2d8475588768dfb7aaf6c3864351d64c