UPSTREAM: HID: wacom: leds: make sure Cintiq 21UX2 and 24HD control the right LEDs
The code for 21UX2 and 24HD makes the LED group 1 on the left, and
the group 0 on the right. The buttons are ordered in the other way,
but libwacom already exports those that way. So we simply can't reassign
LED group 0 to the left buttons, and have to quirk the incoming data...

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Acked-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 5a0fe8abd169be4f2e1f73ad851419d502b5a832)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: I2bf9fdf4a49a4c8042001e23c18d49528e67b300