BACKPORT: drm/i915/guc: add engine mask to GuC client & pass to GuC
The Context Descriptor passed by the kernel to the GuC contains a field
specifying which engine(s) the context will use. Historically, this was
always set to "all of them", but if we had a separate client for each
engine, we could be more precise, and set only the bit for the engine
that the client was associated with. So this patch enables this usage,
in preparation for having multiple clients, though at this point there
is still only a single client used for all supported engines.

Signed-off-by: Dave Gordon <david.s.gordon@intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
(cherry picked from commit e02757d91f7481093742bb32385567499a59354c)

Conflicts:
   drivers/gpu/drm/i915/i915_guc_submission.c: Merge with [4025a0ba9d71
   BACKPORT: drm/i915: Allocate intel_engine_cs structure only for the enabled
   engines] - that backport introduced a custom change which is not present
   in the original patch.

BUG=chromium:965320
TEST=run drm-test/plane_test -f NV12

Change-Id: I421aff2c579967595b185b737600bfce9ecfff74
Signed-off-by: Eugene Nemkov <eugene.nemkov@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1834707
Tested-by: Santosh Kumar Avati <santosh.kumar.avati@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>