UPSTREAM: [media] videodev2.h Add HSV formats
These formats store the color information of the image
in a geometrical representation. The colors are mapped into a
cylinder, where the angle is the HUE, the height is the VALUE
and the distance to the center is the SATURATION. This is a very
useful format for image segmentation algorithms.

Signed-off-by: Ricardo Ribalda Delgado <ricardo.ribalda@gmail.com>
Acked-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit 66b2ab271afc0888b87a44dc946cc68067ba0985)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I2dfc0f5b8a7448d9d6e9663f7fb3eec225500e22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2420361
Reviewed-by: Sean Paul <seanpaul@chromium.org>