UPSTREAM: drm/i915: sseu: Convert subslice count fields to subslice mask
In an upcoming patch we'll need the actual mask of subslices in addition
to their count, so convert the subslice_per_slice field to a mask.
Also we can easily calculate subslice_total from the other fields, so
instead of storing a cached version of this, add a helper to calculate
it.

v2:
- Use hweight8() on u8 typed vars instead of hweight32(). (Ben)

Reviewed-by: Robert Bragg <robert@sixbynine.org> (v1)
Reviewed-by: Ben Widawsky <benjamin.widawsky@intel.com> (v1)
Tested-by: Ben Widawsky <benjamin.widawsky@intel.com> (v1)
Signed-off-by: Imre Deak <imre.deak@intel.com>
(cherry picked from commit 57ec171eae21ee0a2848ade1cb7f8f12a31b0731)

BUG= b:158655609
TEST= boot-up fizz, video playback

Signed-off-by: Cooper Chiou <cooper.chiou@intel.com>
Change-Id: Ifa68feb4cd8c88a8652c4dfc7ffdbb86429cd02b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2971898
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>