UPSTREAM: drm/i915/guc: Keep the execbuf client allocated across reset
In order to avoid some complexity in trying to reconstruct the
workqueues across reset, remember them instead. The issue comes when we
have to handle a reset between request allocation and submission, the
request has reserved space in the wq, but is not in any list so we fail
to restore the reserved space. By keeping the execbuf client intact
across the reset, we also keep the reservations.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20161129121024.22650-5-chris@chris-wilson.co.uk
(cherry picked from commit 4d357af49918e21dde6ecfb9b2f431fe0ab3eff8)

BUG=chromium:965320
TEST=run drm-test/plane_test -f NV12

Change-Id: I765d94191d5980ad16b8ca344905ea571df386ef
Signed-off-by: Eugene Nemkov <eugene.nemkov@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1834986
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Santosh Kumar Avati <santosh.kumar.avati@intel.com>