UPSTREAM: HID: wacom: generic: add support for touchring
Add support for the touchring to the generic code path in support of the second
generation Intuos Pro.

We also add checks for usage->type to ensure that we handle the usage before we
report it, or change the inrange_state based on it.

Signed-off-by: Aaron Skomra <aaron.skomra@wacom.com>
Reviewed-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 60a221869803a0864c4b1683395ef574ee15aba6)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: I4ed57ba716f701d3d5642ac0a474cbd3b3961ca5