UPSTREAM: HID: wacom: generic: support generic touch switch
The second generation Intuos Pro is the first device in the generic codepath
which has a touchswitch. We utilize a flag in wacom_shared in order to report
this switch event received from the pad on the touch input.

Signed-off-by: Aaron Skomra <aaron.skomra@wacom.com>
Reviewed-by: Ping Cheng <pingc@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit d2ec58aee8b1d68d309656a0561c81c12197a987)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: If06934400fa02593b965e1c5cdad544b90a35149