UPSTREAM: kvm: x86: hyperv: avoid livelock in oneshot SynIC timers
If the SynIC timer message delivery fails due to SINT message slot being
busy, there's no point to attempt starting the timer again until we're
notified of the slot being released by the guest (via EOM or EOI).

Even worse, when a oneshot timer fails to deliver its message, its
re-arming with an expiration time in the past leads to immediate retry
of the delivery, and so on, without ever letting the guest vcpu to run
and release the slot, which results in a livelock.

To avoid that, only start the timer when there's no timer message
pending delivery.  When there is, meaning the slot is busy, the
processing will be restarted upon notification from the guest that the
slot is released.

Signed-off-by: Roman Kagan <rkagan@virtuozzo.com>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>

(cherry picked from commit f1ff89ec4447c4e39d275a1ca3de43eed2a92745)

BUG=b:144746965
TEST=Built
Signed-off-by: Matt Delco <delco@chromium.org>

Change-Id: Id8c312caa92cac4134d1d084f7be494361144479
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1927245
Tested-by: Matt Delco <delco@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Matt Delco <delco@chromium.org>