UPSTREAM: HID: wacom: generic: Send BTN_TOOL_PEN in prox once the pen enters range
When a pen is first able to to be sensed by the tablet, we would like
to inform userspace that a tool is nearby so that it can attempt to
perform palm rejection. Unfortunately, we don't know any information
about the tool that is nearby, so the best we can do is send a prox
event for a generic BTN_TOOL_PEN. If the pen later comes closer and
enters proximity, we can determine the actual tool type and send
BTN_TOOL_PEN out of prox if necessary.

Signed-off-by: Ping Cheng <ping.cheng@wacom.com>
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 3e70969e44ee52d72053145dab2cbad74109c685)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: I2926691b565c680df30f0bfea85a821cdcf26db4