BACKPORT: drm/i915: Deprecate I915_SET_COLORKEY_NONE
Deprecate the silly I915_SET_COLORKEY_NONE flag. The obvious
way to disable colorkey is to just set flags to 0, which is
exactly what the intel ddx has been doing all along.

Currently when userspace sets the flags to 0, we end up in a
funny state where colorkey is disabled, but various colorkey
vs. scaling checks still consider colorkey to be enabled, and
thus we don't allow plane scaling to kick in.

In case there is some other userspace out there that actually
uses this flag (unlikely as this is an i915 specific uapi)
we'll keep on accepting it.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20180202204231.27905-1-ville.syrjala@linux.intel.com
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>
(cherry picked from commit 6ec5bd348934887494541bcc9b53d621b1f2962c)

Conflicts:
	drivers/gpu/drm/i915/intel_atomic_plane.c keep DRM_ROTATE_0

BUG=b:149176357
TEST=boot Soraka, run video playback

Signed-off-by: Chiou, Cooper <cooper.chiou@intel.com>
Change-Id: I882f8b3275a9fd105482d43a4f179808e003eb65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2051869
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Grace Kao <grace.kao@intel.com>