UPSTREAM: HID: wacom: Move handling of HID quirks into a dedicated function
We want to keep device-specific quirks as contained as possible so that the
the code remains maintainable. Our 'wacom_setup_device_quirks' function is
the usual place for this, but some quirks need to be applied to the HID
descriptor as it is parsed. This commit introduces a new function which is
called for each usage so that any HID-specific quirks can be applied. The
function now houses quirks that were being done in 'wacom_feature_mapping'
and 'wacom_usage_mapping'.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 578325120ec122db98b3d57e833d16c1bdac6ec6)
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>

Change-Id: Ie89842cb005bf0c5b67d3fb09f13517348132e82