CHROMIUM: iwl7000: separate elements from cfg that are needed by trans_alloc
In order to be able to select the cfg depending on the HW revision or
on the RF ID, we need to set up the trans before selecting the cfg.
To do so, move the elements from cfg that are needed by
iwl_trans_alloc() to a separate struct at the top of the cfg, so it
can be used by other cfg types as well, before selecting the rest of
the configuration.

Change-Id: I816225e03287c2565f8f274edd6206cc8e5676b8
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: d833604ce2e6d6a534a4e565ff084ac0e3243e37