CHROMIUM: staging: gasket: page table: remove extraneous memory barriers
Some explicit memory barriers in the page table code are not necessary,
either because:

(a) The barrier follows a non-relaxed MMIO access that already performs
a read or write memory barrier.

(b) The barrier follows DMA API calls for which the device-visible
effects of IOMMU programming are guaranteed to be flushed to the IOMMU
prior to the call returning, and doesn't need to sync with normal memory
access.

(cherry picked from commit 08d722e741a138beaf6faca4291c3dbdd451b88b
https://darwinn-internal.googlesource.com/platforms/darwinn-kernel-driver/
master)

BUG=b:147826257
TEST=Did not test, intermediate commit in a series of 77 cherry-picks

Signed-off-by: Todd Poynor <toddpoynor@google.com>
Signed-off-by: Richard C Yeh <rcy@google.com>
Change-Id: Id9141b884b6ae39294b5396b23c1743ce50ab6c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2175408
Reviewed-by: Guenter Roeck <groeck@chromium.org>