UPSTREAM: phy: Add MIPI D-PHY configuration options
Now that we have some infrastructure for it, allow the MIPI D-PHY phy's to
be configured through the generic functions through a custom structure
added to the generic union.

The parameters added here are the ones defined in the MIPI D-PHY spec, plus
the number of lanes in use. The current set of parameters should cover all
the potential users.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
(cherry picked from commit 2ed869990e14bc5528aeb00c45e42793c5406637)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Ic5fe0c19ce98f2849baedd8d77c58a3a66f458b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2420042
Reviewed-by: Sean Paul <seanpaul@chromium.org>