UPSTREAM: phy: dphy: Clarify lanes parameter documentation
The lanes parameter is not solely about the number of lanes, but it also
carries the fact that those are the first lanes in use during the
transmission.

It was implicit so far, so make sure it's explicit now.

Suggested-by: Sakari Ailus <sakari.ailus@iki.fi>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
(cherry picked from commit 1baafbe482e54c020751796d7bfdee669acca58b)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Ib7d0a6ebc3b0313d63b2635051e019aebb120a8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2420046
Reviewed-by: Sean Paul <seanpaul@chromium.org>