CHROMIUM: iwl7000: pcie: fix rb_allocator workqueue allocation
We don't handle failures in the rb_allocator workqueue allocation
correctly. To fix that, move the code earlier so the cleanup is
easier and we don't have to undo all the interrupt allocations in
this case.

Change-Id: I486160666c655cbd47e37c8f103b89ef685b6ae1
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 90bb6e7f184b8c547e33b7bc82114f3367d2fb3c