CHROMIUM: iwl7000: mvm: update powersave correctly for D3
This fixes a long-standing bug - we haven't been able to check the
firmware image that was loaded for D3/not-D3 since the introduction
of the unified image...

Fix this by keeping a status flag for D3 instead of checking for
the firmware image that's loaded.

This reduces occurrences of checks for IWL_UCODE_WOWLAN to just the
code that actually loads the image or deals with it in other ways.

Change-Id: If73118faa1c0e9cafcbbeefcce600ac23a36fef1
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: ccbb9db996ec00d4652823bdb5e1a1a1cabde2ba