CHROMIUM: iwl7000: convert the 9260-1x1 device to use the new parameters
The 9260-1x1 device can be differentiated using the PCI device ID.
There is a single occurrence of this device, so continue relying on
the device and subsystem device IDs.

The name of this device was incorrect, so add a new string
specifically for it.

Change-Id: Id27bd9c878b73cb771691cbe6082fd40e079b44d
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 7fe012073f94ed6b2cb685d32f28d13ef9480085