FROMLIST: mfd: cros_ec: query EC protocol version if EC transitions between RO/RW
RO and RW of EC may have different EC protocol version. If EC transitions
between RO and RW, but AP does not reboot (this is true for fingerprint
microcontroller / cros_fp, but not true for main ec / cros_ec), the AP
still uses the protocol version queried before transition, which can
cause problems. In the case of fingerprint microcontroller, this causes
AP to send the wrong version of EC_CMD_GET_NEXT_EVENT to RO in the
interrupt handler, which in turn prevents RO to clear the interrupt
line to AP, in an infinite loop.

Once an EC_HOST_EVENT_INTERFACE_READY is received, we know that there
might have been a transition between RO and RW, so re-query the protocol.

(am from https://lore.kernel.org/patchwork/patch/1154832/)

BUG=chromium:1021692
TEST=Deploy kernel on Nocturne board with its fingerprint microcontroller
	(cros_fp) RO firmware only supports v0 and v1 of EC_CMD_GET_NEXT_EVENT,
	but its RW firmware supports v2 in addition; reboot the fingerprint
	microcontroller but not the AP; observe that RO does not go to
	infinite loop and boots to RW successfully.

Change-Id: I5ad4e5d704cee77da0f05f809947ca3c11be3525
Signed-off-by: Yicheng Li <yichengli@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1903814
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>