HACK: media: rockchip/isp1: Compute plane address with data_offset
With V4L2 MEMORY DMABUF for Multiplane API, it is necessary
to set a file descriptor on each plane. The way to pass an
offset with a DMA-buf is not defined in V4L2 specification.
This is the case as it is needed to specify offsets because
minigbm creates one DMA-buf for all planes. This is hack to
abuse data_offset for the purpose.

BUG=b:140152839, chromium:901264
TEST=Manual, Tast.camera.HAL3Frame, tast.camera.HAL3StillCapture

Change-Id: I833f46316c1736b8102a7d7f5754f29ae0651974
Signed-off-by: Han Lin Chen <hanlinchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1909756
Reviewed-by: Tomasz Figa <tfiga@chromium.org>