HACK: media: vb2: Fill data_offset of vb2_plane by that of v4l2_plane
With V4L2 MEMORY DMABUF for Multiplane API, it is necessary
to set a file descriptor on each plane. The way to pass an
offset with a DMA-buf is not defined in V4L2 specification.
Some drivers need to abuse the data_offset field for the
purpose. This patch fills the data_offset field of vb2_plane
by the same field of v4l2_plane, so drivers may utilize it
when needed.

BUG=b:140152839, chromium:901264
TEST=Manual, Tast.camera.HAL3Frame, tast.camera.HAL3StillCapture

Change-Id: I698be0b5351ebd9c201eb486a7cb977d318754a4
Signed-off-by: Han Lin Chen <hanlinchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1914062
Reviewed-by: Tomasz Figa <tfiga@chromium.org>