CHROMIUM: media: rockchip-vpu: fix MMAP buffers offset on 64-bit
The starting offset for MMAP capture buffers was set to (TASK_SIZE / 2).
However on 64-bit TASK_SIZE is bigger than 32-bits, which is the size of
the offset when passed to user-space. Thus trying to mmap a capture
buffer would result in the driver believing we want to map the
corresponding output buffer.

Fix this by setting the offset to (1UL << 30), which will be valid on
both 32 and 64 bits.

BUG=chromium:1029396
TEST=VEA unittest not failing to mmap on Kevin64.
TEST=VEA unittests passing on Kevin.

Change-Id: Ia047465bdb5addee02c050c511f263ec77a5b59a
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1965748
Reviewed-by: Tomasz Figa <tfiga@chromium.org>